Update Log.cpp
The functions including check point and auto flush log are finished, and the recovery is waiting for the parse of the main. Now it can read log from the disk from check point to the end and parse it to char*, recovering it to the original command, and the rest part just need to do the same as what parse of the main do.